pyproject.toml
[project]
name = "bravobot"
version = "0.1.0"
description = "RAG-based institutional assistant for Institución Universitaria Pascual Bravo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "pgvector>=0.3.6",
    "sentence-transformers>=3.3",
    "torch>=2.5",
    "numpy>=1.26",
    "groq>=0.13",
    "cerebras_cloud_sdk>=1.19",
    "anthropic>=0.39",
    "selectolax>=0.3.25",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
